+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Improved the error messages in **generate-docs**, if an example was not provided.
 * Added to **validate** command a validation that a content entity or a pack name does not contain the words "partner" and "community".
 * Fixed an issue where **update-release-notes** ignores *--text* flag while using *-f*
-* Updated the dockerimage in integration/script yml files to be required.
+* Added a new validation to require the dockerimage key to exist in an integration and script yml files.
 
 # 1.4.5
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.